msi: Support full UI repair installations
This commit addresses an issue that up to now, has been ignored: When
executing `msiexec /i Git-<Version>-<arch>-bit.msi` when Git for
Windows is already installed, this should trigger the Maintenance Welcome
dialog to be shown instead of the standard installation Welcome dialog.
The Maintenance Welcome dialog then progresses to a Maintenance Type
dialog where the user can choose to repair or remove Git for Windows.

The following changes were made in this commit:

* Explicit scheduling of the WelcomeDlg in the InstallUISequence was
  removed from `GitProduct.wxs` to use the scheduling provided by the
  stock WiX WelcomeDlg. The explicit scheduling was only used during
  initial development to focus on the core installation scenarios of
  first-time installations and upgrades.

* An explicit `DialogRef` reference was removed for the
  GitMaintenanceTypeDlg because it is referenced directly by the
  `MaintenanceWelcomeDlg` in the Next button event handler.

* An explicit `DialogRef` reference was created for the
  `MaintenanceWelcomeDlg` because it's a "root" dialog that starts an
  installation UI sequence. This dialog starts the UI installation
  sequence only if the MSI is invoked for installation but Git for Windows
  is already installed.

With this commit, the installer now supports a proper Repair installation
scenario.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>