Make edit-git-bash.exe even smarter
The purpose of edit-git-bash.exe is to modify the string table resources
of `git-bash.exe`. The first string entry of the first string table
resource in `git-bash.exe` is used as the "command line" invocation for
starting the BASH shell for Git for Windows. The standard Git for Windows
distribution allows you to run the BASH shell in one of two terminal
hosts: MinTTY or the standard Windows Command Prompt terminal host
ConHost.exe.

By default, `git-bash.exe` has a string resource that runs the BASH inside
of a MinTTY terminal host. During installation, if the user chooses to run
the BASH shell using Windows ConHost as the terminal host, then
`edit-git-bash.exe` is executed, specifying the path to `git-bash.exe` and
the new command line to use to start the BASH.

The biggest problem is that this was a one-way transformation of
`gti-bash.exe`--once you chose Windows Console Host as the terimnal host,
you can't choose to switch back to MinTTY without re-installing Git for
Windows. This is far from desirable. This is particularly so as this
author attempts to create an MSI-based installation for Git for Windows,
allowing users to change their installation of Git for Windows--including
the user's terminal selection.

This commit changes `edit-git-bash.exe` such that it is capable of
modifying `git-bash.exe`'s string resources back and forth between using
the MinTTY command line invocation and the Windows ConHost command line
invocation as the terminal host for the BASH.

To support this change, instead of passing in the desired command line to
use to invoke the BASH, you instead pass in the desired terminal host (for
the second argument to `edit-git-bash.exe`). This has one major benefit
over passing in the command line: the `git-bash.exe` executable already
contains the correct command lines--if these command lines change over tim
(and they have as recently as August 2016), the installers won't need to
change to modify the string resources appropriately using this program.

It should also be noted that the `git-bash.exe` that is built has two
string resources in the first string table, but only the first string
resource is needed. The first string resource is the command line
invocation to use to start the BASH in MinTTY, while the second string
resource is never read--it is in fact, the string resource to use to start
the BASH using ConHost as the terminal host.

This commit allows these string resources to be "swapped" to effectively
change the active terminal host in use. Furthermore, a second string table
is created with one string whose value is the current active terminal
host. Upon initially running this program, there is no string table 2, and
so one can assume this program has never modified `git-bash.exe`, and
therefore, the default terimnal host is active, MinTTY. Otherwise, string
table 2's first string contains one of MinTTY or ConHost.

The InnoSetup `install.iss` file has been updated to pass the correct
arguments to `edit-git-bash.exe` and the function defined for carrying out
the user's selected option has also had its name updated to better reflect
what it's doing.

Lastly, some MSI-specific files have been updated. Previously, TERMINAL
would be set to one of MinTTY or CmdPrompt (as that's what's output in the
current `install_options.txt` file). But in the InnoSetup installer, the
two option values are MinTTY or ConHost. Therefore, the MSI has been
updated accordingly.

**UPDATE**
I have since learned from @dscho that the reason both command lines are
present in `git-bash.exe` is so that if, for some reason, mintty.exe could
not be located or executed, then the wrapper that starts `git-bash.exe`
would "fall back" to using the Windows Console Host (i.e. the second entry
in the first string table). Since I put so much work into rewriting how
this program works, if it should ever become useful, I still wanted to
have this branch and commit around. But I have since been given a
suggestion that may completely eliminate the need for even having
`edit-git-bash.exe` at all.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>