Make edit-git-bash.exe smarter
The purpose of edit-git-bash.exe is to modify the string table resources
of `git-bash.exe`. The first string entry of the first string table
resource in `git-bash.exe` is used as the "command line" invocation for
starting the BASH shell for Git for Windows. The standard Git for Windows
distribution allows you to run the BASH shell in one of two terminal
hosts: MinTTY or the standard Windows Command Prompt terminal host
ConHost.exe.

By default, `git-bash.exe` has a string resource that runs the BASH inside
of a MinTTY terminal host. During installation, if the user chooses to run
the BASH shell using Windows ConHost as the terminal host, then
`edit-git-bash.exe` is executed, specifying the path to `git-bash.exe` and
the new command line to use to start the BASH.

The biggest problem is that this was a one-way transformation of
`gti-bash.exe`--once you chose Windows Console Host as the terimnal host,
you can't choose to switch back to MinTTY without re-installing Git for
Windows. This is far from desirable. This is particularly so as this
author attempts to create an MSI-based installation for Git for Windows,
allowing users to change their installation of Git for Windows--including
the user's terminal selection.

This commit changes `edit-git-bash.exe` such that it is capable of
modifying `git-bash.exe`'s string resources back and forth between using
the MinTTY command line invocation and the Windows ConHost command line
invocation as the terminal host for the BASH.

To support this change, it made sense to internalize the command line
strings inside of `edit-git-bash.exe`. So, instead of passing in a new
command line for the second argument to the program, you would now pass in
one of `MinTTY` or `ConHost` (these strings are detected
case-insensitively). The usage text for `edit-git-bash.exe` has also been
updated accordingly. Also, I'm using a bit outdated copy of
`git-bash.exe` locally, and so I went looking to see if the string
resources have changed. It appears that
`git-for-windows/MINGW-packages/blob/master/mingw-w64-git/git-bash.rc` had
a recent change (commit 636d775) to update the string resources for the
newer version of MinTTY, so they've been updated accordingly.

It should also be noted that the `git-bash.exe` that is built has two
string resources in the first string table, but only the first string
resource is needed. The first string resource is the command line
invocation to use to start the BASH in MinTTY, while the second string
resource is never read--it is in fact, the string resource to use to start
the BASH using ConHost as the terminal host. (Perhaps in the future, it
would be possible to specify a parameter to git-bash.exe to choose the
string resource to use for starting the BASH, completely eliminating the
need for `edit-git-bash.exe` altogether?) So, when modifying the string
table resource, `edit-git-bash.exe` only creates one string entry in the
string table, namely the command line invocation to use to start
`git-bash.exe` in the user's terminal host of choice.

The InnoSetup `install.iss` file has been updated to pass the correct
arguments to `edit-git-bash.exe` and the function defined for carrying out
the user's selected option has also had its name updated to better reflect
what it's doing.

Lastly, some MSI-specific files have been updated. Previously, TERMINAL
would be set to one of MinTTY or CmdPrompt (as that's what's output in the
current `install_options.txt` file). But in the InnoSetup installer, the
two option values are MinTTY or ConHost. Therefore, the MSI has been
updated accordingly.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>